Gestion erreurs et leurs affichages

+++ www/Views/showPages.php
@@ -10,12 +10,12 @@
         </div>
 
         <p>Description: <?= nl2br(htmlspecialchars($page["description"])) ?></p>
-        <?php if ($page['user_id'] == $_SESSION['id']): ?>
+        <?php if (isset($_SESSION['id']) && $page['user_id'] == $_SESSION['id']): ?>
             <a href="/page/edit?id=<?= $page['id'] ?>">Éditer</a> |
             <a href="/page/delete?id=<?= $page['id'] ?>"
             onclick="return confirm('Supprimer cette page ?')">
             Supprimer
-            </a>
+            </a>            
         <?php endif; ?>
         <a href="/<?= htmlspecialchars($page["slug"]) ?>">Voir la page</a>
     </div>
